Refactor: Moves files out of component Main pages on pages folder API urls on the api folder
