+++ src/App.tsx
@@ -6,12 +6,12 @@ import 'bootstrap/dist/js/bootstrap.bundle.min';
 import { UserContext } from "./components/userContext";
 
 import NavBar from './components/navbar';
-import Home from './components/home';
+import Home from './pages/home';
 import CreateUser from './components/createUser';
-import BugList from './components/bugList';
+import BugList from './pages/bugs/bugList';
 import ViewBug from './components/viewBug';
 import CreateBug from './components/createBug';
-import ViewUser from "./components/viewUser";
+import ViewUser from "./pages/users/viewUser";
 
 const App: React.FC = () => {
     const stored: String | null = sessionStorage.token ? sessionStorage.token : null;
