+++ src/components/createUser.tsx
@@ -5,8 +5,8 @@ import { Company as CompanySchema, User as UserSchema } from "../interfaces";
 import { UserContext } from "./userContext";
 import Spinner from "./utilities/spinner";
 import ErrorText from "./utilities/errorText";
-import { apiUserBase, login } from "./api/userApi";
-import { apiCompanyBase } from "./api/companyAPI";
+import { apiUserBase, login } from "../api/userApi";
+import { apiCompanyBase } from "../api/companyAPI";
 
 // if newUser prop is true this form will register a user, else for login 
 const CreateUser: React.FC = () => {
